Add files via upload
For the begining of you job, you can download fleetco from the github.
The way to work this project with PHP7 is to replace /include/cipheredes.php by the modificated file (by Enrique Barchiesi) in the attachement. It is not full secure login but it works.
You should backup the original file.
You can change the password for admin user (or any user) by phpmyadmin (password in the table `carrierusers`) to “44v2v294y2v2” (that means the password you must fill in during login will be 123456), later (after login) you can change this pasword.