dompdf
To start working the PDF by dompdf you can download actual version of dompdf from github.
Than you can delete the old one version of dompdf in the path /plugins/dompdf and substitue it with the new one.
Than you can delete the old one version of the page2pdf.php in the path /plugins and substitute it with the one below.
It is quite working - you can generate the pdf. The only bug that I can not solve is missing table borders.